+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Creation of autoware-compatible lanelet2 maps
 - Set custom plot limits when adding scenario or aerial image
+- Map German traffic sign 252 to sign to 260
 
 ### Changed
 - Versions of third-party packages